simulation of a pulse with absorbing boundary conditions

+++ fd1d/cuda/test_1_2.cu
@@ -15,7 +15,7 @@
 
 __device__
 float gaussian(int t, int t0, float sigma) {
-    return exp(-0.5*(t - t0)/sigma*(t - t0)/sigma);
+    return expf(-0.5f*(t - t0)/sigma*(t - t0)/sigma);
 }
 
 
@@ -23,7 +23,7 @@ __global__
 void exfield(int t, int nx, float *ex, float *hy) {
     /* calculate the Ex field */
     for (int i = idx+1; i < nx; i += stx)
-        ex[i] += 0.5 * (hy[i-1] - hy[i]);
+        ex[i] += 0.5f * (hy[i-1] - hy[i]);
     /* put a Gaussian pulse in the middle */
     if (idx == nx/2) ex[nx/2] = gaussian(t, 40, 12.0f);
 }
@@ -36,7 +36,7 @@ void hyfield(int nx, float *ex, float *hy, float *bc) {
     if (idx == nx-1) ex[nx-1] = bc[3], bc[3] = bc[2], bc[2] = ex[nx-2];
     /* calculate the Hy field */
     for (int i = idx; i < nx-1; i += stx)
-        hy[i] += 0.5 * (ex[i] - ex[i+1]);
+        hy[i] += 0.5f * (ex[i] - ex[i+1]);
 }
